+++ nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.infra/src/main/java/nts/uk/ctx/at/shared/infra/entity/statutory/worktime_new/company/KshstComFlexSetPK.java
@@ -29,7 +29,7 @@ public class KshstComFlexSetPK implements Serializable {
     
     /** The year. */
     @Column(name = "YEAR")
-    private short year;
+    private int year;
 
     /**
      * Instantiates a new kshst com flex set PK.
