+++ nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.infra/src/main/java/nts/uk/ctx/at/shared/infra/entity/statutory/worktime_new/share/KshstRegLaborTime.java
@@ -32,7 +32,7 @@ public abstract class KshstRegLaborTime extends UkJpaEntity {
 
 	/** The week str. */
 	@Column(name = "WEEK_STR")
-	private short weekStr;
+	private int weekStr;
 
 	/** The daily time. */
 	@Column(name = "DAILY_TIME")
